Major UI changes to the maestro, performer, and listener screens Added ability to see participants on the screen
